Move documenter to seperate file
